requirements.txt
azure-devops>=7.1.0b3
msrest>=0.7.1
python-dotenv>=1.0.0
pydantic>=2.6.1
requests>=2.31.0
aiohttp>=3.8.0
pandas>=1.5.0
jira>=3.5.0
tqdm>=4.66.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0 